requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
Flask==0.10.1
ipdb==0.9.3
ipython==4.2.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
simplegeneric==0.8.1
traitlets==4.2.1
Werkzeug==0.11.8
